Adds failsafe options and updates README
Implements failsafe options in the settings screen to help users recover from app misbehavior by providing ways to stop all audio or force exit the application.

Updates the README to include a disclaimer that the app has only been tested on Ubuntu 24.04.3.
Also, runs the AppImage in the background so that the terminal is not blocked.

Increments the version number.